Fixing undefined index notices in locale_field_entity_form_submit()
The unset of form fields is causing a notice from
locale_field_entity_form_submit() men locale.module is enabled.

This commit fixes it by setting the field acces to FALSE instead. It is
therefore also needed to set required = FALSE on the borger.dk URL
field.

Changing all form boolean fields to FALSE instead of 0 which is not
correct according to the official Drupal coding standards.